Make draw extend LayerGroupPointer
